refactor: simplify and parameterize crawler cron-job
