+++ bundestag.io/dip/tsconfig.json
@@ -1,4 +1,5 @@
 {
+  "exclude" : ["src/**/*.spec.ts", "jest.config.ts"],
   "compilerOptions": {
     /* Visit https://aka.ms/tsconfig.json to read more about this file */
