--- services/cron-jobs/crawler/.local.env
@@ -1,2 +1,0 @@
- MONGO_DB_URL=mongodb://localhost:27017/bundestagio
- DIP_API_KEY=N64VhW8.yChkBUIJeosGojQ7CSR2xwLf3Qy7Apw464
