+++ services/cron-jobs/crawler/package.json
@@ -8,15 +8,15 @@
   "license": "MIT",
   "scripts": {
     "build": "tsc",
-    "dev:import-procedures": "ts-node-dev -r dotenv/config ./src/import-procedures dotenv_config_path=./.local.env",
+    "dev:import-procedures": "ts-node-dev -r dotenv/config ./src/import-procedures",
     "start:import-procedures": "node ./build/import-procedures.js"
   },
   "dependencies": {
     "@democracy-deutschland/bundestag.io-definitions": "^1.0.2",
     "@democracy-deutschland/bundestagio-common": "^0.1.13",
-    "cross-fetch": "^3.1.4",
     "dotenv": "^10.0.0",
-    "winston": "^3.3.3"
+    "graphql": "^15.5.0",
+    "graphql-request": "^3.4.0"
   },
   "devDependencies": {
     "ts-node-dev": "^1.1.6"
